+++ NEWS.md
@@ -1,3 +1,9 @@
+# infer 0.3.0.9000
+
+- Stop using package {assertive} in favor of custom type checks (#149)
+- Fixed `t_stat()` to use `...` so `var.equal` works
+- With the help of @echasnovski, fixed `var.equal = TRUE` for `specify() %>% calculate(stat = "t")`
+
 # infer 0.3.0
 
 - Added `conf_int` logical argument and `conf_level` argument to `t_test()`